refactor(#1869): move #1869 window tests into renamed big_promoted_seek_tests (window_builder submodule); apply cargo fmt
Avoids growing the over-threshold data_access/mod.rs (epic #1116): register the
tests via the already-declared sibling module (renamed from big_promoted_crc_tests
to big_promoted_seek_tests to reflect its expanded scope), a net-zero mod.rs edit.
